Move `InMemorySessionStore` in it's own module implementation
