+++ samples/minimal/build.gradle.kts
@@ -49,6 +49,7 @@ dependencies {
     implementation(libs.androidx.activity.compose)
     implementation(projects.libraries.matrix.api)
     implementation(projects.libraries.matrix.impl)
+    implementation(projects.libraries.sessionStorage.implMemory)
     implementation(projects.libraries.designsystem)
     implementation(projects.libraries.architecture)
     implementation(projects.libraries.core)
