+++ settings.gradle.kts
@@ -69,3 +69,4 @@ include(":samples:minimal")
 include(":libraries:encrypted-db")
 include(":libraries:session-storage:api")
 include(":libraries:session-storage:impl")
+include(":libraries:session-storage:impl-memory")
